Implement generic encoder implementation and swerve module IO that uses an absolute encoder directly connected to the turn Spark MAX
We're testing many different swerve module configurations, so these changes make it easier to swap out the hardware on a few modules at once and view the differences.

Ideally, we wouldn't use the turn motor gear ratio for PID control on the new "spark max with encoder" module IO implementation at all, but I have it configured in that way for now so we can directly compare different module types without re-tuning the PIDs and potentially breaking the established module control logic.